+++ includes/jgl_draw_primitive.h
@@ -16,8 +16,8 @@ void draw_triangle_texture(Pixel *points, Uv *uvs, float alpha, size_t nb = 1);
 void fill_triangle_color(Pixel *points, Color *colors, size_t nb = 1);
 
 vector<Vector2> calc_point(int width);
-void draw_line(Vector2 p1, Vector2 p2, int width, Color p_color, c_viewport *viewport = nullptr);
-void draw_point(Vector2 center, int width, Color p_color, c_viewport *viewport = nullptr);
+void draw_line(Vector2 p1, Vector2 p2, size_t width, Color p_color, c_viewport *viewport = nullptr);
+void draw_point(Vector2 center, size_t width, Color p_color, c_viewport *viewport = nullptr);
 void draw_rectangle(Vector2 p_tl, Vector2 p_tr, Vector2 p_dl, Vector2 p_dr, Color p_color, c_viewport *viewport = nullptr);
 void fill_rectangle(Vector2 p_tl, Vector2 p_tr, Vector2 p_dl, Vector2 p_dr, Color p_color, c_viewport *viewport = nullptr);
 void draw_rectangle(Vector2 pos, Vector2 size, Color p_color, c_viewport *viewport = nullptr);
